refactor(core): move binding processors to binding package
Leave public annotations in current package to avoid breaking changes
