+++ springwolf-core/src/main/java/io/github/stavshamir/springwolf/asyncapi/scanners/bindings/ProcessedOperationBinding.java
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: Apache-2.0
-package io.github.stavshamir.springwolf.asyncapi.scanners.channels.operationdata;
+package io.github.stavshamir.springwolf.asyncapi.scanners.bindings;
 
 import com.asyncapi.v2.binding.operation.OperationBinding;
 import lombok.Data;
